Move reporting of metrics from SV to C++

+++ cosim/dut_if/rvfi/rvfi.cpp
@@ -206,7 +206,7 @@ void rvfi::enter_debug_mode(rv_instr_t& instr) {
   if (!FLAGS_cosim)
     return;
 
-  if (instr.pc.pc_rdata == FLAGS_debug_entry_pc) {
+  if ((int)instr.pc.pc_rdata == FLAGS_debug_entry_pc) {
 
     rv_debug_t debug;
     std::cout <<" Enter Debug Mode debug_entry_pc :"<<std::hex<<FLAGS_debug_entry_pc<<"\n";
@@ -226,7 +226,7 @@ void rvfi::exit_debug_mode(rv_instr_t& instr) {
   if (!FLAGS_cosim)
     return;
 
-  if (instr.pc.pc_rdata == FLAGS_debug_exit_pc) {
+  if ((int)instr.pc.pc_rdata == FLAGS_debug_exit_pc) {
 
     rv_debug_t debug;
     std::cout <<" Exit Debug Mode debug_exit_pc :"<<std::hex<<FLAGS_debug_exit_pc<<"\n";
